`--input-url` CLI flag, better credential handling
You can now pass a URL to a .tar.gz archive, which will be downladed by
either of the two processes in `inputs.nf`.

This archive must contain 3 files in the top-level directory:

- `data.csv`: Timeseries data
- `metadata.json`: Metadata
- `rejects.csv`: A rejects file

These files will be used as outputs to either of the two processes. This
makes it easy to do runs with custom data.

---

Credentials: the following credentials are passed as Nextflow Secrets:

- COVIDESTIM_DBSTAN_USER
- COVIDESTIM_DBSTAN_PASS
- COVIDESTIM_DBSTAN_DBNAME
- COVIDESTIM_DBSTAN_HOST
- COVIDESTIM_JWT

COVIDESTIM_JWT should always be set. The other variables should be set
to the empty string if not used. The final variable used for API/DB
operations, COVIDESTIM_ENDPOINT, is managed in `nextflow.config` by the
"db_local" and "db_prod" profiles.